refactor: collapse global barrel into global/index.ts
